==> notebooks/1.1_data_wrangling_training.ipynb <==
# Data Wranging -> Preparing Training Data

Our training data comes nicely to us from the spider database files 'train_spider.json' and 'train_others.json'.

The first was built by the team that prepared this data and the second is a collection based on common public datasets.

Included in the files are training questions with the corresponding schema_id, correct query to answer the question, and some parsed out data for the question and query.

I could actually pull the tokenized words from the question, but I want that practice so I'll do that. I also haven't thought of a good use-case to pull in the query information (I wouldn't know how to train on it), so I'm only going to pull the questions and the corresponding schema_id.
