==> notebooks/2.0_processing_vector_db.ipynb <==
# Saving Schema Data to Vector Database

My initial thought was to build a model that checks for similarity between the prompt and the schema information. That may still be part of the long-term roadmap, but in doing research I found I could get acceptable results in a much simpler way using a vector database through langchain. By savings documents related to our schemas to this database, we can then query based on the question and return the documents that are most closely related. With this, we'll try to all be bundled into langchain. woohoo!

Big shoutout to this great blogpost that provided some of the framework: https://canvasapp.com/blog/text-to-sql-in-production

from watermark import watermark
print(watermark())

import pandas as pd
import json
import os

from dotenv import load_dotenv

from langchain import SQLDatabase
from langchain.docstore.document import Document
from langchain.vectorstores import Chroma
from langchain.embeddings import HuggingFaceEmbeddings

from sqlalchemy import exc

## Current Method

In this first launch, I settled on saving the schema, table, column list, DDL, and 3 sample rows into each document.
There is also metadata attached to each document with the schema, table, and column list to easily pull from the query results.

## Establish Vector Database

There are a few options for databases, but I'll go with Chroma because it is open source, makes local device use "easy", and has built-in connections with langchain.

A key component in running apps using langchain is the ability store and work with embeddings, which is how AI models natively represent data of all kinds. Langchain will provide the application framework and Chroma will provide the vector store.

Within that we can also do some information retrieval from the database, finding the most relevant tables based on the user question. This will allow us to engineer a better prompt to feed to our gpt chatbot.

#setup embeddings using HuggingFace
embeddings  = HuggingFaceEmbeddings()

### Create Functions

def get_json(path):
    """Return json file from specified filepath"""
    with open(path, "r") as f:
        json_file = json.load(f)
    
    return json_file

#point to where our .sqlite files are saved
db_path = '../data/processed/db/'

def connect_db(db_path, target_schema):
    """
    Take in the identified schema and connect to the sqlite database with that name
    """
    db_filepath = db_path
    db_filename = target_schema + '.sqlite'

    #point to database
    base_dir = os.path.dirname(os.path.abspath(db_filepath+db_filename)) #get the full path within the device
    db_path = os.path.join(base_dir, db_filename) #combine with filename to get db_path
    db = SQLDatabase.from_uri("sqlite:///" + db_path) #connect via the lanchain method

    return db

#define new document builder
def prep_chroma_documents_v2(json_path):
    """Take json file and work through it to prepare for load to Chroma.
    Instead of list comprehension, establis the blank list and loop through the json.
    Then saves to the list using the langchain docstore.document -> Document module.

    This version - adding the table info using the langchain SQLDatabase SQLAlchemy wrapper to get table info to add to metadata.
    Would like to not reconnect to the database each time, but instead connect to each schema once and then loop through the tables. But I think this will be easier for now, even if it's less efficient.

    This works specifically with the content and metadata we want for this project"""
    docs = []
    for item in get_json(json_path):
        #connect to database
        db = connect_db(db_path=db_path, target_schema=item['schema'])

        #create variables
        schema = item['schema']
        table = item['table']
        columns = json.dumps([col['c_name'] for col in item['columns']])
        try:
            table_info = db.get_table_info_no_throw(table_names=[table]) #put try-except here becasue there are some issues in the source sqlite database. I want to call this out, but continue.
        except exc.SQLAlchemyError as e:
            table_info = ""
            print(schema + "-" + table + ": " + str(e))
        except TypeError as te:
            print(schema + "-" + table + ": " + str(te))          

        #create document
        doc = Document(
            page_content=f"""
Schema: {schema}
Table: {table}
Columns: {columns}
DDL:
    {table_info}
""",
            metadata={
                'schema': schema,
                'table': table,
                'columns': columns,
            }
        )
        docs.append(doc)
    
    return docs

table_docs = prep_chroma_documents_v2('../data/interim/schema_info.json')

### Create VectorDB

#setup directory to store database on disk
persist_dir_new = '../data/processed/chromadb/schema-table-info'

###commenting out since this was already run and created###

#vectordb_new = Chroma.from_documents(documents=table_docs, embedding=embeddings, persist_directory=persist_dir_new)

#vectordb_new.persist() #should only need to run this out of the notebook, not a .py file.
vectordb_new=None

### Test - Table Info Method

# load from disk
vectordb_info = Chroma(persist_directory=persist_dir_new, embedding_function=embeddings)

test = "What is the average number of employees of the departments whose rank is between 10 and 15?" #one of the prompts from the training data

test_docs = vectordb_info.similarity_search(test, k=3)

test_docs

#test mmr retriever mentioned in langchain docs (Maximum Marginal Relevance)
retriever = vectordb_info.as_retriever(search_type="mmr")

retriever.get_relevant_documents(test)

This didn't put it in the 1st spot, but it is in the top three. After testing on all of the questions, this document structure ended up being the best performing of the three methods I tried. This is despite the fact that some of the table info steps didn't run due to database issues for 3-4 of the tables.

## Other Methods

### Simple Schema-Table

In this first iteration I tried, the document content only had the schema-table combo.

def prep_chroma_documents(json_path):
    """Take json file and work through it to prepare for load to Chroma.
    Instead of list comprehension, establis the blank list and loop through the json.
    Then saves to the list using the langchain docstore.document -> Document modeul

    This works specifically with the content and metadata we want for this project"""
    docs = []
    for item in get_json(json_path):
        doc = Document(
            page_content=f"Schema Table: {item['schema_split'] + ' ' + item['table_split']}",
            metadata={
                'schema': item['schema'],
                'table': item['table'],
                'columns': json.dumps([col['c_name'] for col in item['columns']])
            }
        )
        docs.append(doc)
    
    return docs

table_info = prep_chroma_documents('../data/interim/schema_info.json')

table_info[:3]

#setup directory to store database on disk
persist_dir = '../data/processed/chromadb/schema-table-split'

###commenting out since we already created this database

#vectordb = Chroma.from_documents(documents=table_info, embedding=embeddings, persist_directory=persist_dir)

#### Persist DB

#vectordb.persist() #think I need to call this mainly because I'm in an ipynb.
vectordb=None

## Test Vector Database Query

# load from disk
vectordb = Chroma(persist_directory=persist_dir, embedding_function=embeddings)

query = "How many heads of the departments are older than 56?" #one of the prompts from the training data

docs = vectordb.similarity_search(query, k=1)

print('Most Likely Schema and Table:\n')
i=0
for doc in docs:
    print(docs[i].metadata['schema'] + ' - ' + docs[i].metadata['table'])
    i+=1

# Test Entity Extraction

https://canvasapp.com/blog/text-to-sql-in-production

This blog suggests doing some entity extraction from the question. I can try the same method they do of calling to a chat LLM. I could also try using a HuggingFace model specific to text classification or entity extraction.

I could try to train an entity recognition model on the schema information. But first I'll try using a chat model.

from dotenv import load_dotenv

from langchain import HuggingFaceHub
from langchain.prompts import PromptTemplate
from langchain.chains import LLMChain

#establish examples of what we want to get back from the model

extract_prompt = """question: How many teams are there?
response: team

question: What user spent the most money in March?
response: user, money

question: What is the name of the instructor who advises the student with the greatest number of total credits?
response: instructor, advisor, student, credits
"""

#create instruction for the LLM

instruction = """
For a given question, determine the keywords for determining what tables should be used to write a SQL query to answer the question.
"""

#create the full prompt content combining the examples and the instruction

prompt = (
    instruction
    + "\nHere are are three question/response examples: "
    + extract_prompt
    + "\n\nquestion: "
    + test
    + "\nresponse:"
)

#structure the prompt in the langchain template

entity_extr_prompt = PromptTemplate(
    input_variables=[],
    template = prompt
)

#establish the LLM - using google flan again

#get api key
load_dotenv()
hf_api_token = os.getenv('hf_token')

#add path to HF repo
repo_id = 'tiiuae/falcon-7b-instruct'

#establish llm model
llm = HuggingFaceHub(repo_id=repo_id, huggingfacehub_api_token=hf_api_token, model_kwargs={"temperature": .005, "max_length": 512})

#create an LLM chain
chain = LLMChain(llm=llm, prompt=entity_extr_prompt, verbose=False)

#test on question
results= chain.predict() #run llm
print(results)

#run the resul throug the similarity search
vectordb_info.similarity_search(results, k=3)

This single test is promising as it's the first to return our target schema as the primary result. I'm worried about how to efficiently test this, since each run requires a call to an LLM. Maybe I can go to my validation page and just test on a sample of the questions.

### Follow-Up:

I was able to test on 50 of the random questions - only getting 40% of them correctly matched. Trying to go for a larger sample size ran into the hourly limits of Hugging Face Inference. Due to the poor performance on the small sample size and inability to test more, I won't go with this method.

## Final Test -> Remove Table Info from Docs

This will just make the docs have the schema - table - columns to see if that improves the overall results. If not, I'll go with my v2.

#define new document builder
def prep_chroma_documents_v3(json_path):
    """Take json file and work through it to prepare for load to Chroma.
    Instead of list comprehension, establis the blank list and loop through the json.
    Then saves to the list using the langchain docstore.document -> Document module.

    This version - adding the table info using the langchain SQLDatabase SQLAlchemy wrapper to get table info to add to metadata.
    Would like to not reconnect to the database each time, but instead connect to each schema once and then loop through the tables. But I think this will be easier for now, even if it's less efficient.

    This works specifically with the content and metadata we want for this project"""
    docs = []
    for item in get_json(json_path):
        #create variables
        schema = item['schema']
        table = item['table']
        columns = json.dumps([col['c_name'] for col in item['columns']])      

        #create document
        doc = Document(
            page_content=f"""
Schema: {schema}
Table: {table}
Columns: {columns}
""",
            metadata={
                'schema': schema,
                'table': table,
                'columns': columns,
            }
        )
        docs.append(doc)
    
    return docs

#create new docs
new_docs = prep_chroma_documents_v3('../data/interim/schema_info.json')

new_docs[:3]

#setup directory to store database on disk
persist_dir_3 = '../data/processed/chromadb/sch-tab-col'

###commenting out as this was already created

#vectordb_3 = Chroma.from_documents(documents=new_docs, embedding=embeddings, persist_directory=persist_dir_3)

#vectordb_3.persist() #think I need to call this mainly because I'm in an ipynb.
vectordb_3=None

#### Test New DB

# load from disk
vectordb_3 = Chroma(persist_directory=persist_dir_3, embedding_function=embeddings)

test = "What is the average number of employees of the departments whose rank is between 10 and 15?" #one of the prompts from the training data

test_docs = vectordb_3.similarity_search(test, k=3)

test_docs

**Observations:**
Looks good on the test, I'll go to my validation notebook and try it on all of the questions.

### Follow-up:

I did the test on the full question list and it returned slightly worse results than the db with the table info. I'll stick with that one.
