==> notebooks/3.0_simple_sql_chain.ipynb <==
# Langchain SQL Chain

My takeaway from a few weeks of learning langchain is that it's very cool. In this notebook, I test the "out-of-the-box" SQL Database Chain.

This takes a specified llm model, database, and a prompt template (which the langchain peeps have created as a standard) and runs through a chain. This chain feeds in the table info from the tables in the database and prompts the LLM to answer an input quesiton using that info. It writes a SQL, runs it against the database, and then answers the question for you.

For my workflow, I don't know which database to use yet. That's were the vector database we created comes in. Using a similarity search, I want to pull down the most likely schema and table and save the metadata as variables. Then will take the standard langchain prompt template and tweak it slightly to enter the model. With this simple first test, I'll make it a specific chain with only one table required to answer the question, whch I'll feed to to model directly.

from watermark import watermark
print(watermark())

## Imports

import pandas as pd
import os
from dotenv import load_dotenv

from langchain.vectorstores import Chroma
from langchain.embeddings import HuggingFaceEmbeddings
from langchain import HuggingFaceHub

from langchain import SQLDatabase, SQLDatabaseChain
from langchain.prompts.prompt import PromptTemplate

#set this to false to get rid of warning
os.environ["TOKENIZERS_PARALLELISM"] = "false"

## Load Top Results from VectorDB

We'll load this from the "persist directory" we saved in the last step.

#setup embeddings using HuggingFace and the directory location
embeddings  = HuggingFaceEmbeddings()
persist_dir = '../data/processed/chromadb/schema-table-split'

# load from disk
vectordb = Chroma(persist_directory=persist_dir, embedding_function=embeddings)

#run prompt as query and get most likely results
query = "How many heads of the departments are older than 56?" #first prompt from the training data

result = vectordb.similarity_search(query, k=1)

result

## Save Metadata to Variables

top_schema = result[0].metadata['schema']
top_table = result[0].metadata['table']
table_cols = result[0].metadata['columns']

print(top_schema, top_table, table_cols)

## Build SQL Agent Through HF Hub

The docs on langchain are focused around either accessing the OpenAI API or a local model. I'm hoping I can mix both and access a Hugging Face model through an API without having to store it locally. I think I can do this through the Hugging Face Hub, by combined the steps in different parts of the documentation:
- [Langchain - Hugging Face Hub](https://python.langchain.com/docs/modules/model_io/models/llms/integrations/huggingface_hub)
- [Langchain - SQL Agents](https://python.langchain.com/docs/modules/chains/popular/sqlite)

My original plan was to use Open AI's chatGPT model, but HuggingFace has their [Infererence API](https://huggingface.co/inference-api) which allows free semi-restricted access to some great LLMs. But you still need an API Token, which you can find [here](https://huggingface.co/settings/tokens).

In initial testing I used the falcon-7b-instruct model to varying performance. But I've tested some others as well. On the inference API, you are a little limited in the models you can choose. The 40 or 70b models are too large, but would likely have better performance. So the trick is finding ways to use the smaller models at no cost, but still get good results. I ended up using the google-flan-xxl model mainly which I've updated this notebook to use as well.

### Load API Token

load_dotenv()
hf_api_token = os.getenv('hf_token')

#add path to HF repo
repo_id = 'tiiuae/falcon-7b-instruct'

### Point to LLM Model

#establish llm model
llm = HuggingFaceHub(repo_id=repo_id, huggingfacehub_api_token=hf_api_token, model_kwargs={"temperature": 0.5, "max_length": 512})

### Setup Simple SQL Agent
Langchain has some awesome features, but I'll start with a simple chain that points to one table.

First, I need to use the output of my vector db query to point to the sqlite database we want to work with.

#### Point to sqlite db location

Using this method expands the location to give me the full directory path.

In this test - I didn't pass our metdata from the similarity search and just kept it simple and pointed to the db. I'll work on that later.

BASE_DIR = os.path.dirname(os.path.abspath('../data/processed/db/department_management.sqlite'))
db_path = os.path.join(BASE_DIR, "department_management.sqlite")

db_path

#### Establish db

db = SQLDatabase.from_uri("sqlite:///" + db_path)

#### Create Prompt Template
I'll even specify the single table here to really spoon feed the model. But this general framework for the template comes directly from langchain.

sql_agent_prompt_template = """You are an expert data analyst. Given an input question, first create a syntactically correct {dialect} query to run, then look at the results of the query and describe the answer.
Use the following format:

Question: Question here
SQLQuery: SQL Query to run
SQLResult: Result of the SQLQuery
Answer: Final answer here

Only use the following tables:
{table_info}

Question: {input}"""

sql_prompt = PromptTemplate(
    input_variables=["input", "table_info", "dialect"], template=sql_agent_prompt_template)

#### Setup db_chain

db_chain = SQLDatabaseChain.from_llm(llm, db, prompt=sql_prompt, verbose=True)

#### Test with our simple question

db_chain.run(sql_prompt.format(input="How many heads of the departments are older than 56?", table_info="head", dialect="sqlite"))

**Observations:**
This simple chain method is working, although not as cleanly as I would like. The chain is inconsistent in returning "5" as the answer vs describing it with something like "There are 5 department heads older than 56."

Or sometimes it just doesn't fill in with a complete statement. But by and large this is working when it is very specific.

## Create Functions
Even if I want some cleanup, this technically works which is great. But it is very segmented. So I want to pull the steps together so I don't have to manually change variables in each step. This will test loading in all table info (not just the one we know to be correct) This will also make future testing easier.

def db_select(persist_dir, query):
    """
    Load the schema info vector database from disk and run the input question against it to return the most likely database we need to pull data from.
    """
    #setup embeddings using HuggingFace and the directory location
    embeddings  = HuggingFaceEmbeddings()
    per_dir = persist_dir

    # load from disk
    vectordb = Chroma(persist_directory=per_dir, embedding_function=embeddings)

    #run prompt as query and get most likely results
    result = vectordb.similarity_search(query, k=1)

    #save variables
    top_schema = result[0].metadata['schema']
    top_table = result[0].metadata['table']
    table_cols = result[0].metadata['columns']

    top_result = (top_schema, top_table, table_cols)

    return top_result

def locate_and_connect_db(filepath, filename):
    """Locate the absolute path of the given SQLITE database and connect to it via the langchain SQLDatabase.from_uri method.
    filepath is the filepath within the repo, example '../data/db/data.db'
    filename is just the filename.filetype, example 'data.db'
    """
    base_dir = os.path.dirname(os.path.abspath(filepath+filename)) #get the full path within the device
    db_path = os.path.join(base_dir, filename) #combine with filename to get db_path
    db = SQLDatabase.from_uri("sqlite:///" + db_path) #connect via the lanchain method

    return db

def load_llm_model(env_variable, repo_id='google/flan-t5-xxl', temp=0.2, max_length=512):
    """
    Take in the target hugging face repo and the api key to setup the llm to use in our query chain
    env_variable is the name of the variable that stores your API key in your .env file
    """
    load_dotenv()
    hf_api_token = os.getenv(env_variable)
    llm = HuggingFaceHub(repo_id=repo_id, huggingfacehub_api_token=hf_api_token, model_kwargs={"temperature": temp, "max_length": max_length})

    return llm

###update the prompt template to a more detailed one I found suggested in a langchain issue.

template = """You are a SQLite expert. Given an input question, first create a syntactically correct SQLite query to run, then look at the results of the query and explain the results in context of the original question.
            Unless the user specifies in the question a specific number of examples to obtain, query for at most {top_k} results using the LIMIT clause as per SQLite. You can order the results to return the most informative data in the database.
            Query only the columns that are needed to answer the question. Wrap each column name in double quotes (") to denote them as delimited identifiers.
            Pay attention to use only the column names you can see in the tables below. Be careful to not query for columns that do not exist. Also, pay attention to which column is in which table.
            Pay attention to use date('now') function to get the current date, if the question involves "today".

            Use the following format:

            Question: Question here
            SQLQuery: SQL Query to run
            SQLResult: Result of the SQLQuery
            Answer: Final answer here

            Only use the following tables:
            {table_info}

            Question: {input}"""


def create_prompt_template(prompt_template=template, input_vars=['input', 'table_info', 'top_k']):
    prompt_temp = PromptTemplate(input_variables=input_vars, template=prompt_template) #create our prompt template
    
    return prompt_temp

def create_sql_chain(llm, db, prompt_template, verbose=True):
    """Take in prompt template and input variables, along with the llm and db created from other functions to create our SQL Chain"""
    db_chain = SQLDatabaseChain.from_llm(llm, db, prompt=prompt_template, verbose=verbose, return_intermediate_steps=True) #create our database chain
    
    return db_chain

# def run_chain(db_chain, question, table, sql_dialect='sqlite'):
#     """Function to run chain that will end our overall application function by taking in your question and variables created by the other functions"""
#     db_chain.run(sql_prompt.format(input=question, table_info=table, dialect=sql_dialect))

def sql_analyst(question, vector_db_path='../data/processed/chromadb/schema-table-split', db_root_path='../data/processed/db/', sql_dialect='sqlite', env_api_key_var='hf_token'):
    """take in user info on where the databases and the api key are located and answer their question using the sql chain."""
    
    #query our vector database with the question to get the schema most related to the question.
    top_result = db_select(persist_dir=vector_db_path, query=question)
    top_schema = top_result[0]
    top_table = top_result[1]

    #use this result to establish our database
    db = locate_and_connect_db(filepath=db_root_path, filename=top_schema+'.'+sql_dialect)

    #initialize our large language model - needs an API key - we'll use the standard variables
    llm = load_llm_model(env_variable=env_api_key_var)

    #create prompt template using our preset template
    prompt_template = create_prompt_template(prompt_template=template)

    #setup the sql chain using the standard variables
    sql_chain = create_sql_chain(db=db, llm=llm, prompt_template=prompt_template)

    #run sql_chain on question
    print(top_schema)
    print(prompt_template)
    sql_chain(question)

#### Test

sql_analyst("How many heads of the departments are older than 56?")

This continues to be inconsistent at best. In this case it created a working query, but the wrong one. Part of this could be the capability of the model itself. But even with some testing on chat GPT 3.5 online, it doesn't always work to just load in ALL of the table information. So we need to try to make the prompt better and maybe even guide the model through some steps. So next I will explore the SQL Database Sequential Chain.
